Python/requirements.txt
python-dotenv == 1.0.1
azure-cognitiveservices-speech==1.34.0
anyio==4.2.0
elevenlabs==0.2.8
google-cloud-texttospeech == 2.16.3
keyboard==0.13.5
mutagen==1.46.0
openai==1.7.2
pydantic==1.10.13
pygame-ce==2.4.0
rich==13.7.0
soundfile==0.12.1
tiktoken==0.5.1
pyserial==3.5
